% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_mrt.R
\name{convert_mrt}
\alias{convert_mrt}
\title{Make a mrt_table object}
\usage{
convert_mrt(gen, count, censor_col = NULL)
}
\arguments{
\item{data}{Data frame containing mrt assay data. Generation number must be in the first column, and the remaining columns must contain percentage of viable plates for a given strain at the specified generations.}
}
\description{
Creates a mrt_table object from a data frame containing mrt assay data.
}
