% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_mrt_table.R
\name{make_mrt_table}
\alias{make_mrt_table}
\title{Make a mrt_table object}
\usage{
make_mrt_table(x)
}
\arguments{
\item{data}{Data frame containing mrt assay data. Generation number must be in the first column, and the remaining columns must contain percentage of viable plates for a given strain at the specified generations.}
}
\description{
Creates a mrt_table object from a data frame containing mrt assay data.
}
\examples{
mrt_plot()
}

