% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrt_plot.R
\name{mrt_plot}
\alias{mrt_plot}
\title{Create mrt plot}
\usage{
mrt_plot(data)
}
\arguments{
\item{data}{mrt_table object created using make_mrt_table}
}
\description{
Creates a survival curve plot from a mrt_table object.
}
\examples{
mrt_plot()
}

