% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_logranks.R
\name{perform_logranks}
\alias{perform_logranks}
\title{Perform pairwise logrank comparisons between samples in a mrt_table object}
\usage{
perform_logranks(data, correct = FALSE)
}
\arguments{
\item{data}{mrt_data object}

\item{correct}{apply bonferroni correction to the p-values}

\item{exp}{data contains an "exp" with experiment/replicate ID which should be added as a covariate}
}
\description{
Perform pairwise logrank comparisons between samples in a mrt_table object
}
