% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_mrt.R
\name{convert_mrt}
\alias{convert_mrt}
\title{Make a mrt_table object from generation, count and censoring data for one sample}
\usage{
convert_mrt(gen, count, censor_col = NULL)
}
\arguments{
\item{gen}{vector of generations}

\item{count}{vector of counts (number of individuals still alive) at each generation}

\item{censor_col}{optional vector of number of censored individuals at each generation}
}
\description{
Creates a mrt
}
