% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_mrt_table.R
\name{make_mrt_table}
\alias{make_mrt_table}
\title{Make a mrt_table object from a dataframe containing survival data for one or more samples}
\usage{
make_mrt_table(data, censored = NULL)
}
\arguments{
\item{data}{Data frame containing mrt assay data. Generation number must be in the first column, and the remaining columns must contain the number of viable plates for a given strain at the specified generations.}

\item{censored}{Optional data frame containing number of censored individuals at each generation, with the same column order as data.}
}
\description{
Creates a mrt_table object from a data frame containing mrt assay data.
}
